Move PrettyInfo into ValueInfo
This is one more step towards fixing orphan instances.
