+++ src/Feldspar/Core/Middleend/FromTyped.hs
@@ -58,7 +58,7 @@ import Feldspar.Core.Reify (ASTF, unASTF, render)
 import Feldspar.Core.Types (TypeRep(..), defaultSize, TypeF(..), (:>)(..))
 import qualified Feldspar.Core.Types as T
 import Feldspar.Core.UntypedRepresentation hiding (Type(..), ScalarType(..))
-import Feldspar.Core.ValueInfo (ValueInfo(..), prettyVI)
+import Feldspar.Core.ValueInfo (ValueInfo(..), PrettyInfo(..))
 import Feldspar.Range (Range(..))
 import qualified Feldspar.Core.Representation as R
 import Feldspar.Core.Representation (AExpr((:&)), Expr((:@)))
@@ -530,17 +530,6 @@ data FrontendPass
      | FPUnAnnotate
      | FPCreateTasks
      deriving (Eq, Show, Enum, Bounded, Read)
-
--- | Overloaded pretty printing of annotations (for instance range information)
-
-class PrettyInfo a where
-  prettyInfo :: U.Type -> a -> String
-
-instance PrettyInfo ValueInfo where
-  prettyInfo = prettyVI
-
-instance PrettyInfo () where
-  prettyInfo _ _ = ""
 
 instance Pretty UntypedFeld where
   pretty = pretty . annotate (const ())
